moved character into to character info component
